Support remote linked images: filter & download
Books to building will embed remote linked images downloaded from
internet. The links embedded would need to be modified to match the
path the asset added while rendering the document. So some info about
these assets should be prepared ahead of rendering. One is original
link and the other is hashed filename that saves the remote content.
And when the book is rendering, the path of remote linked images are
patched with hashed one that assets can be found correctly while
reading.

For downloading assets, `ureq` is chosen without introducing too much
dependencies. And mockall for dev dependency to test without a server.